+++ spec/dummy/client/app/startup/serverRegistration.jsx
@@ -1,4 +1,4 @@
-import 'babel-polyfill';
+import '@babel/polyfill';
 import 'es5-shim';
 
 // Shows the mapping from the exported object to the name used by the server rendering.