+++ spec/dummy_no_webpacker/client/webpack.client.js
@@ -4,7 +4,7 @@ module.exports = {
   entry: [
     'es5-shim/es5-shim', // for poltergeist
     'es5-shim/es5-sham', // for poltergeist
-    'babel-polyfill',
+    '@babel/polyfill',
     'jquery',
     'jquery-ujs',
     'startup/clientRegistration',